requirements.txt
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.2 